Start editing the css of articles
changed the css of infobox, hr between title and content and put
display none on mw-editsection as that feature won’t be available here
for now.

Also removed the hard limit of 2000px on article height, formerly used
to have all articles with at least that width, no longer needed.